requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
jiter==0.9.0
mem0ai==0.1.77
monotonic==1.6
numpy==2.2.4
openai==1.69.0
portalocker==2.10.1
posthog==3.23.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pydantic==2.11.0
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
pytz==2024.2
qdrant-client==1.13.3
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
